Move NodeParameters and some utility functions.
