+++ src/signer_node/mod.rs
@@ -2,6 +2,12 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+mod utils;
+mod node_parameters;
+
+use utils::sender_index;
+pub use node_parameters::NodeParameters;
+
 use std::collections::BTreeMap;
 use std::convert::TryInto;
 use std::sync::mpsc::{channel, Receiver, Sender, TryRecvError};
@@ -115,14 +121,6 @@ pub enum NodeState {
         block_shared_keys: Option<(bool, FE, GE)>,
         candidate_block: Option<Block>,
     },
-}
-
-fn sender_index(sender_id: &SignerID, pubkey_list: &[PublicKey]) -> usize {
-    //Unknown sender is already ignored.
-    pubkey_list
-        .iter()
-        .position(|pk| pk == &sender_id.pubkey)
-        .unwrap()
 }
 
 impl<T: TapyrusApi, C: ConnectionManager> SignerNode<T, C> {
@@ -886,51 +884,6 @@ impl<T: TapyrusApi, C: ConnectionManager> SignerNode<T, C> {
         Parameters {
             threshold: t,
             share_count: n.clone(),
-        }
-    }
-}
-
-pub struct NodeParameters<T: TapyrusApi> {
-    pub pubkey_list: Vec<PublicKey>,
-    pub threshold: u8,
-    pub private_key: PrivateKey,
-    pub rpc: std::sync::Arc<T>,
-    pub address: Address,
-    pub signer_id: SignerID,
-    pub self_node_index: usize,
-    pub round_duration: u64,
-    pub skip_waiting_ibd: bool,
-}
-
-impl<T: TapyrusApi> NodeParameters<T> {
-    pub fn new(
-        to_address: Address,
-        pubkey_list: Vec<PublicKey>,
-        private_key: PrivateKey,
-        threshold: u8,
-        rpc: T,
-        round_duration: u64,
-        skip_waiting_ibd: bool,
-    ) -> NodeParameters<T> {
-        let secp = secp256k1::Secp256k1::new();
-        let self_pubkey = private_key.public_key(&secp);
-        let signer_id = SignerID {
-            pubkey: self_pubkey,
-        };
-
-        let mut pubkey_list = pubkey_list;
-        &pubkey_list.sort();
-        let self_node_index = sender_index(&signer_id, &pubkey_list);
-        NodeParameters {
-            pubkey_list,
-            threshold,
-            private_key,
-            rpc: Arc::new(rpc),
-            address: to_address,
-            signer_id,
-            self_node_index,
-            round_duration,
-            skip_waiting_ibd,
         }
     }
 }
